# ACCServer/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "shadowDemo restful API.",
    "version": "1.0"
  },
  "host": "localhost:8030",
  "basePath": "/acc",
  "paths": {
    "/aPIAccessReqLog": {
      "get": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessReqLog"
        ],
        "summary": "详情;",
        "operationId": "getAPIAccessReqLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getAPIAccessReqLogResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "put": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessReqLog"
        ],
        "summary": "修改;",
        "operationId": "updateAPIAccessReqLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessReqLog"
        ],
        "summary": "创建;",
        "operationId": "createAPIAccessReqLog",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessReqLog"
        ],
        "summary": "删除;",
        "operationId": "deleteAPIAccessReqLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/aPIAccessReqLogs": {
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessReqLog"
        ],
        "summary": "查询列表;",
        "operationId": "searchAPIAccessReqLog",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CreateEndTime": {
                  "description": "创建截止时间",
                  "type": "string",
                  "format": "date-time"
                },
                "CreateStartTime": {
                  "description": "创建开始时间",
                  "type": "string",
                  "format": "date-time"
                },
                "IDS": {
                  "description": "id",
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "PageNum": {
                  "type": "integer",
                  "format": "int64"
                },
                "PageSize": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/searchAPIAccessReqLogResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/aPIAccessResLog": {
      "get": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessResLog"
        ],
        "summary": "详情;",
        "operationId": "getAPIAccessResLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getAPIAccessResLogResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "put": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessResLog"
        ],
        "summary": "修改;",
        "operationId": "updateAPIAccessResLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessResLog"
        ],
        "summary": "创建;",
        "operationId": "createAPIAccessResLog",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessResLog"
        ],
        "summary": "删除;",
        "operationId": "deleteAPIAccessResLog",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/aPIAccessResLogs": {
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "aPIAccessResLog"
        ],
        "summary": "查询列表;",
        "operationId": "searchAPIAccessResLog",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CreateEndTime": {
                  "description": "创建截止时间",
                  "type": "string",
                  "format": "date-time"
                },
                "CreateStartTime": {
                  "description": "创建开始时间",
                  "type": "string",
                  "format": "date-time"
                },
                "IDS": {
                  "description": "id",
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "PageNum": {
                  "type": "integer",
                  "format": "int64"
                },
                "PageSize": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/searchAPIAccessResLogResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "APIAccessReqLog": {
      "type": "object",
      "properties": {
        "APIURL": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedBy": {
          "type": "string"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "int64"
        },
        "IPAddress": {
          "type": "string"
        },
        "MerchantName": {
          "type": "string"
        },
        "MerchantNo": {
          "type": "string"
        },
        "Method": {
          "type": "string"
        },
        "OrderNo": {
          "type": "string"
        },
        "Request": {
          "type": "string"
        },
        "RequestNo": {
          "type": "string"
        },
        "RequestTime": {
          "type": "string",
          "format": "date-time"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "shadowDemo/model/do"
    },
    "APIAccessResLog": {
      "type": "object",
      "properties": {
        "APIURL": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedBy": {
          "type": "string"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "int64"
        },
        "IPAddress": {
          "type": "string"
        },
        "MerchantName": {
          "type": "string"
        },
        "MerchantNo": {
          "type": "string"
        },
        "Method": {
          "type": "string"
        },
        "OrderNo": {
          "type": "string"
        },
        "Request": {
          "type": "string"
        },
        "RequestNo": {
          "type": "string"
        },
        "RequestTime": {
          "type": "string",
          "format": "date-time"
        },
        "Response": {
          "type": "string"
        },
        "ResponseTime": {
          "type": "string",
          "format": "date-time"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "shadowDemo/model/do"
    },
...
    "GenericMessageBody": {
      "type": "object",
      "properties": {
        "Code": {
          "type": "integer",
          "format": "int64"
        },
        "Msg": {
          "type": "string"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
    "GetAPIAccessReqLogResponseBody": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/APIAccessReqLog"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
    "GetAPIAccessResLogResponseBody": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/APIAccessResLog"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
    "SearchAPIAccessReqLogResponseBody": {
      "type": "object",
      "properties": {
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "Result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/APIAccessReqLog"
          }
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
    "SearchAPIAccessResLogResponseBody": {
      "type": "object",
      "properties": {
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "Result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/APIAccessResLog"
          }
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
  },
  "responses": {
...
    "genericError": {
      "description": "错误消息",
      "schema": {
        "$ref": "#/definitions/GenericMessageBody"
      }
    },
    "genericSuccess": {
      "description": "成功消息",
      "schema": {
        "$ref": "#/definitions/GenericMessageBody"
      }
    },
    "getAPIAccessReqLogResponse": {
      "description": "详情出参",
      "schema": {
        "$ref": "#/definitions/GetAPIAccessReqLogResponseBody"
      }
    },
    "getAPIAccessResLogResponse": {
      "description": "详情出参",
      "schema": {
        "$ref": "#/definitions/GetAPIAccessResLogResponseBody"
      }
    },
...
    "searchAPIAccessReqLogResponse": {
      "description": "查询列表",
      "schema": {
        "$ref": "#/definitions/SearchAPIAccessReqLogResponseBody"
      }
    },
    "searchAPIAccessResLogResponse": {
      "description": "查询列表",
      "schema": {
        "$ref": "#/definitions/SearchAPIAccessResLogResponseBody"
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}